Change implicit casts to explicit.
No functional changes.

+++ libntp/clocktime.c
@@ -80,7 +80,7 @@ clocktime(
 	 */
 	if (*yearstart) {
 		/* -- get time stamp of potential solution */
-		test[0] = (uint32_t)(*yearstart) + tmp;
+		test[0] = (uint32_t)(*yearstart) + (uint)tmp;
 		/* -- calc absolute difference to receive time */
 		diff[0] = test[0] - rec_ui;
 		if (diff[0] >= 0x80000000u)
@@ -102,12 +102,12 @@ clocktime(
 	 * around the guess and select the entry with the minimum
 	 * absolute difference to the receive time stamp.
 	 */
-	y = ntp_to_year(rec_ui - tmp);
+	y = ntp_to_year(rec_ui - (uint)tmp);
 	for (idx = 0; idx < 3; idx++) {
 		/* -- get year start of potential solution */
 		ystt[idx] = year_to_ntp(y + idx - 1);
 		/* -- get time stamp of potential solution */
-		test[idx] = ystt[idx] + tmp;
+		test[idx] = ystt[idx] + (uint)tmp;
 		/* -- calc absolute difference to receive time */
 		diff[idx] = test[idx] - rec_ui;
 		if (diff[idx] >= 0x80000000u)
@@ -143,6 +143,6 @@ year_to_ntp(
 	int32_t year)
 {
 	uint32_t days;
-	days = ntpcal_days_in_years(year-1) - DAY_NTP_STARTS + 1;
+	days = (uint32_t)ntpcal_days_in_years(year-1) - DAY_NTP_STARTS + 1;
 	return days * SECSPERDAY;
 }
